backend/requirements.txt
flask
flask-cors
torch
torchvision
pillow
numpy
timm
gunicorn